fix: modify fetch color location (#158) color 데이터를 fetch 하는 위치 수정 부모 -> colorList
